Create different podspec file path for watchOS
