+++ Library/shared/Core/ColorResource.swift
