+++ Library/shared/Core/FileResource.swift
