+++ Library/shared/Core/FontResource.swift
