+++ Library/shared/Foundation/NSBundle+FileResource.swift
