+++ Library/shared/Foundation/NSData+FileResource.swift
