+++ rswift.library/rswift.library.h
@@ -1,0 +1,19 @@
+//
+//  rswift.library.h
+//  rswift.library
+//
+//  Created by Tomas Harkema on 14-05-16.
+//  Copyright © 2016 Mathijs Kadijk. All rights reserved.
+//
+
+#import <WatchKit/WatchKit.h>
+
+//! Project version number for rswift.library.
+FOUNDATION_EXPORT double rswift_libraryVersionNumber;
+
+//! Project version string for rswift.library.
+FOUNDATION_EXPORT const unsigned char rswift_libraryVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <rswift_library/PublicHeader.h>
+
+
